chore(tests): put most tests into the same file
Due to us including the raw disk images within the tests files and having so many of them, they occupied (when compiled) about 1.5-2GBs. I tried to make this somewhat more moderate by putting them back to the same file, but within their respective modules for organizational purposes
