--- tests/remove_dir.rs
@@ -1,57 +1,0 @@
-mod common;
-use common::*;
-
-use rstest::*;
-use rstest_reuse::*;
-use test_log::test;
-
-#[test]
-#[apply(fs)]
-
-fn remove_empty_dir(fs: FileSystem<MemoryDevice, DefaultClock>) {
-    let dir_path = "/emptydir/";
-
-    fs.remove_empty_dir(dir_path).unwrap();
-
-    // the directory should now be gone
-    let dir_result = fs.read_dir(dir_path);
-    match dir_result {
-        Err(err) => match err {
-            FSError::NotFound => (),
-            _ => panic!("unexpected IOError: {err:?}"),
-        },
-        _ => panic!("the directory should have been deleted by now"),
-    }
-}
-
-#[test]
-#[apply(fs)]
-
-fn remove_nonempty_dir_with_readonly_file(fs: FileSystem<MemoryDevice, DefaultClock>) {
-    fs.show_hidden(true);
-
-    let dir_path = "/rootdir/";
-
-    // the directory should contain a read-only file (hidden.txt)
-    let del_result = fs.remove_dir_all(dir_path);
-    match del_result {
-        Err(err) => match err {
-            FSError::ReadOnlyFile => (),
-            _ => panic!("unexpected IOError: {err:?}"),
-        },
-        _ => panic!("the directory shouldn't have been removed already"),
-    }
-
-    // this should now remove the directory
-    fs.remove_dir_all_unchecked(dir_path).unwrap();
-
-    // the directory should now be gone
-    let dir_result = fs.read_dir(dir_path);
-    match dir_result {
-        Err(err) => match err {
-            FSError::NotFound => (),
-            _ => panic!("unexpected IOError: {err:?}"),
-        },
-        _ => panic!("the directory should have been deleted by now"),
-    }
-}
